requirements.txt
playwright>=1.47
python-dotenv>=1.0
requests>=2.31
pandas>=2.2
